refactor: create custom hook for the scrollableTab logic
